refactor: rename for scope
