+++ src/sputnik/src/js/apis/node/text_encoding/mod.rs
@@ -2,7 +2,7 @@ mod javy;
 
 use crate::errors::js::JUNO_SPUTNIK_ERROR_JS_API_TEXT_ENCODING;
 use crate::js::apis::node::text_encoding::javy::polyfill::register;
-use crate::js::js_utils::throw_js_exception;
+use crate::js::inner_utils::throw_js_exception;
 use rquickjs::{Ctx, Error as JsError};
 
 // ===========================================================================================
